requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.30
asyncpg>=0.29.0
redis>=5.0.4
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
aiofiles>=23.2.1
pydantic>=2.7.1
pydantic-settings>=2.2.1
dashscope>=1.16.1
faster-whisper>=1.0.2
edge-tts>=6.1.9
qiniu>=7.15.0
python-dotenv>=1.0.1